// src/app/modules/checklist/checklist.interface.ts
// models/Checklist.ts

export interface Question {
  question: string;
  answer: string;
  explanation?: string;
}

export interface IChecklist {
  orderId: string;
  userId: string;
  questions: Question[];
  createdAt?: string; // Optional, will be auto-generated by MongoDB
  updatedAt?: string; // Optional, will be auto-generated by MongoDB
}